# api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Timestamp Converter API",
    "description": "API for converting timestamps between formats and timezones with support for batch processing, custom formatting, and visualization.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "url": "https://tsconv.com/support",
      "email": "support@tsconv.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://tsconv.com/api",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
...
  "paths": {
    "/convert": {
      "get": {
        "tags": ["conversion"],
        "summary": "Convert a timestamp or date string",
        "description": "Convert between Unix timestamps and human-readable dates with optional timezone conversion and custom formatting.",
        "operationId": "convertTimestamp",
        "parameters": [
          {
            "name": "timestamp",
            "in": "query",
            "description": "Unix timestamp in seconds",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "date",
            "in": "query",
            "description": "ISO 8601 date string",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "format",
            "in": "query",
            "description": "Custom format name",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "timezone",
            "in": "query",
            "description": "Source timezone",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "targetTimezone",
            "in": "query",
            "description": "Target timezone for conversion",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeFormats",
            "in": "query",
            "description": "Include all available formats",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful conversion",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/timezone": {
      "get": {
        "tags": ["timezone"],
        "summary": "Get timezone information and perform conversions",
        "description": "Get information about a specific timezone or convert between timezones.",
        "operationId": "getTimezoneInfo",
        "parameters": [
          {
            "name": "timezone",
            "in": "query",
            "description": "Target timezone",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "timestamp",
            "in": "query",
            "description": "Reference timestamp",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "includeTransitions",
            "in": "query",
            "description": "Include DST transitions",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful timezone information retrieval",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimezoneResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Timezone not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ConversionResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "input": {
                "type": "string",
                "example": "1640995200"
              },
              "timestamp": {
                "type": "number",
                "example": 1640995200
              },
              "formats": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                },
                "example": {
                  "iso8601": "2022-01-01T00:00:00.000Z",
                  "utc": "Sat, 01 Jan 2022 00:00:00 GMT",
                  "local": "12/31/2021, 7:00:00 PM",
                  "custom": "12/31/2021 07:00 PM",
                  "relative": "2 years ago"
                }
              }
            }
          },
          "metadata": {
            "type": "object",
            "properties": {
              "processingTime": {
                "type": "number",
                "example": 25
              },
              "cache": {
                "type": "object",
                "properties": {
                  "hit": {
                    "type": "boolean",
                    "example": false
                  },
                  "ttl": {
                    "type": "number",
                    "example": 300
                  }
                }
              }
            }
          }
        }
      },
...
      "TimezoneResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "identifier": {
                "type": "string",
                "example": "America/New_York"
              },
              "displayName": {
                "type": "string",
                "example": "Eastern Time (US & Canada)"
              },
              "currentOffset": {
                "type": "number",
                "example": -300
              },
              "isDST": {
                "type": "boolean",
                "example": false
              },
              "region": {
                "type": "string",
                "example": "America"
              },
              "countryCode": {
                "type": "string",
                "example": "US"
              },
              "aliases": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": ["EST", "EDT"]
              }
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "BAD_REQUEST"
              },
              "message": {
                "type": "string",
                "example": "Invalid timestamp format"
              },
              "details": {
                "type": "object",
                "properties": {
                  "suggestions": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": ["Use Unix timestamp format"]
                  }
                }
              },
              "timestamp": {
                "type": "string",
                "example": "2024-01-15T12:00:00.000Z"
              }
            }
          }
        }
      }
    }
  }
}